Update clearoutside to work with polars
