+++ pvcast/weather/clearoutside.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import datetime as dt
 import logging
 from dataclasses import InitVar, dataclass, field
 from urllib.parse import urljoin
@@ -21,11 +22,9 @@ class WeatherAPIClearOutside(WeatherAPI):
     sourcetype: str = field(default="clearoutside")
     url: str = field(init=False)
     _url_base: InitVar[str] = field(default="https://clearoutside.com/forecast/")
-    _columns: list[str] = field(default_factory=list)
 
     def __post_init__(self, _url_base: str) -> None:
         self.url = self._url_formatter(_url_base)
-        self._columns = ["cloud_coverage", "wind_speed", "temperature", "humidity"]
 
     def _url_formatter(self, url_base: str) -> str:
         """Format the url to the API."""
@@ -41,8 +40,6 @@ class WeatherAPIClearOutside(WeatherAPI):
     def _process_data(self) -> pl.DataFrame:
         """Process weather data scraped from the clear outside website.
 
-        Credits to https://github.com/davidusb-geek/emhass for the parsing code.
-
         This function takes no arguments, but response.content must be retrieved from self._raw_data.
         """
         # raw response data from request
@@ -51,37 +48,36 @@ class WeatherAPIClearOutside(WeatherAPI):
         response = self._raw_data
 
         # response (source) data bucket
-        weather_df = pl.DataFrame(index=self.source_dates, columns=self._columns)
+        weather_df = pl.DataFrame()
+        datetimes = self.source_dates
+        n_days = int(self.max_forecast_days / dt.timedelta(days=1))
 
-        # parse the data
-        n_days = int(self.max_forecast_days / pl.Timedelta(days=1))
+        # Parse HTML content once
+        soup = BeautifulSoup(response.content, "lxml")
+
         for day_int in range(n_days):
-            result = BeautifulSoup(response.content, "html.parser").find_all(
-                id=f"day_{day_int}"
-            )
+            # find the table for the day
+            result = soup.select(f"#day_{day_int}")
             if len(result) != 1:
                 _LOGGER.warning("No data for day %s.", day_int)
                 break
 
-            table = result[0]
-
             # find the elements in the table
+            table = result[0]
             data = self._find_elements(table)
+            data = data.with_columns(datetimes[day_int * 24 : (day_int + 1) * 24])
 
             # insert the data into the source data bucket
-            weather_df.iloc[day_int * 24 : (day_int + 1) * 24] = data
+            weather_df = weather_df.vstack(data)
 
-        # check that all rows with NaN are at the end of the data
-        rows_with_nan = weather_df.isna().any(axis=1)
-        if not rows_with_nan.sum() == 0:
-            _LOGGER.debug("Dropping %s rows with NaN.", rows_with_nan.sum())
-            if not weather_df.isna().any(axis=1).diff().sum() == 1:
-                _LOGGER.warning("Found NaN in the middle of the data.")
-                weather_df.interpolate(
-                    method="linear", inplace=True, limit_area="inside"
-                )
-            weather_df.dropna(inplace=True)
+        # check NaN values distribution
+        nan_vals = weather_df.with_columns(pl.all().is_null().cast(int).diff().sum())
+        if any(col.item() > 1 for col in nan_vals[0]):
+            raise ValueError(f"Found more than one intermediate NaN value: {nan_vals}")
 
+        # interpolate NaN values
+        weather_df = weather_df.interpolate()
+        weather_df = weather_df.drop_nulls()
         return weather_df
 
     def _find_elements(self, table: BeautifulSoup) -> pl.DataFrame:
@@ -100,33 +96,22 @@ class WeatherAPIClearOutside(WeatherAPI):
         list_tables = [list_tables[i] for i in sel_cols]
 
         # building the raw DF container
-        raw_data = pl.DataFrame(index=range(24), columns=col_names, dtype=float)
+        raw_data = pl.DataFrame({"index": range(24)})
         for count_col, col in enumerate(col_names):
             list_rows = list_tables[count_col].find_all("li")
-            for count_row, row in enumerate(list_rows):
-                raw_data.loc[count_row, col] = float(row.get_text())
 
-        # select subset of columns
-        raw_data = raw_data[
-            [
-                "Total Clouds (% Sky Obscured)",
-                "Wind Speed/Direction (mph)",
-                "Temperature (°C)",
-                "Relative Humidity (%)",
-            ]
-        ]
+            # create empty column and fill it with data up len(column_data)
+            column_data = pl.Series(col, [float(row.get_text()) for row in list_rows])
+            raw_data = pl.concat([raw_data, column_data.to_frame()], how="horizontal")
 
         # rename columns
-        raw_data.rename(
-            columns={
-                "Total Clouds (% Sky Obscured)": "cloud_coverage",
-                "Wind Speed/Direction (mph)": "wind_speed",
-                "Temperature (°C)": "temperature",
-                "Relative Humidity (%)": "humidity",
-            },
-            inplace=True,
+        raw_data = raw_data.select(
+            pl.col("Total Clouds (% Sky Obscured)").alias("cloud_coverage"),
+            pl.col("Wind Speed/Direction (mph)").alias("wind_speed"),
+            pl.col("Temperature (°C)").alias("temperature"),
+            pl.col("Relative Humidity (%)").alias("humidity"),
         )
 
         # convert wind speed to m/s
-        raw_data["wind_speed"] = raw_data["wind_speed"] * 0.44704
+        raw_data = raw_data.with_columns(pl.col("wind_speed") * 0.44704)
         return raw_data
