+++ pvcast/weather/weather.py
@@ -17,7 +17,7 @@ from pvlib.location import Location
 from requests import Response
 from voluptuous import All, Datetime, Optional, Range, Required, Schema
 
-from ..model.util import _timedelta_to_pl_duration
+from ..util.timestamps import timedelta_to_pl_duration
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -42,38 +42,6 @@ WEATHER_SCHEMA = Schema(
         ],
     }
 )
-
-# temperature conversion functions ("°C", "°F", "C", "F")
-TEMP_CONV_DICT = {
-    "F": {
-        "C": lambda x: (5 / 9) * (x - 32),
-    },
-    "C": {
-        "C": lambda x: x,
-    },
-}
-
-# speed conversion functions ("m/s", "km/h", "mi/h", "ft/s", "kn")
-SPEED_CONV_DICT = {
-    "m/s": {
-        "km/h": lambda x: x * 3.6,
-    },
-    "km/h": {
-        "m/s": lambda x: x / 3.6,
-    },
-    "mi/h": {
-        "m/s": lambda x: x / 2.23694,
-    },
-    "ft/s": {
-        "m/s": lambda x: x / 3.28084,
-    },
-    "kn": {
-        "m/s": lambda x: x / 1.94384,
-    },
-}
-
-# combine temperature and speed conversion dictionaries
-CONV_DICT = {**TEMP_CONV_DICT, **SPEED_CONV_DICT}
 
 
 @dataclass
@@ -142,7 +110,7 @@ class WeatherAPI(ABC):
         return self.start_forecast + self.max_forecast_days - self.freq_source
 
     @property
-    def source_dates(self) -> pl.datetime_range:
+    def source_dates(self) -> pl.Series:
         """
         Get the pl.DatetimeIndex to store the forecast. These are only used if missing from API, the weather API can
         also return datetime strings and in that case this index is not needed and even not preferred.
@@ -153,40 +121,15 @@ class WeatherAPI(ABC):
 
     @staticmethod
     def get_source_dates(
-        start: dt.datetime, end: dt.datetime, int: dt.timedelta
-    ) -> pl.datetime_range:
+        start: dt.datetime, end: dt.datetime, inter: dt.timedelta
+    ) -> pl.Series:
         """
         Get the pl.DatetimeIndex to store the forecast. These are only used if missing from API, the weather API can
         also return datetime strings and in that case this index is not needed and even not preferred.
         """
-        return pl.datetime_range(start, end, interval=int, timezone=dt.timezone.utc)
-
-    @staticmethod
-    def convert_unit(data: pl.Series, from_unit: str, to_unit: str) -> pl.Series:
-        """Convert units of a pl.Series.
-
-        :param data: The data to convert. This should be a pl.Series.
-        :param to_unit: The unit to convert to.
-        :return: Data with applied unit conversion.
-        """
-        if not isinstance(data, pl.Series):
-            raise TypeError("Data must be a pl.Series.")
-
-        # remove degree symbol from units if present
-        from_unit = from_unit.replace("°", "")
-        to_unit = to_unit.replace("°", "")
-
-        if from_unit not in CONV_DICT:
-            raise ValueError(f"Conversion from unit [{from_unit}] not supported.")
-        if from_unit == to_unit:
-            return data
-        if to_unit not in CONV_DICT[from_unit]:
-            raise ValueError(
-                f"Conversion from [{from_unit}] to [{to_unit}] not supported."
-            )
-
-        # do unit conversion
-        return CONV_DICT[from_unit][to_unit](data)
+        return pl.datetime_range(
+            start, end, inter, time_zone=str(dt.timezone.utc), eager=True
+        )
 
     @abstractmethod
     def _process_data(self) -> pl.DataFrame:
@@ -265,7 +208,7 @@ class WeatherAPI(ABC):
         try:
             data_dict = {
                 "source": self.__class__.__name__,
-                "interval": _timedelta_to_pl_duration(self.freq_source),
+                "interval": timedelta_to_pl_duration(self.freq_source),
                 "data": data_dict,
             }
             WEATHER_SCHEMA(data_dict)
