+++ tests/unit/weather/test_weather_platform.py
@@ -1,10 +1,12 @@
 """Test all configured weather platforms that inherit from WeatherAPI class."""
 from __future__ import annotations
 
+import datetime as dt
 from typing import Any, Generator
 from urllib.parse import urljoin
 
-import pandas as pd
+# import pandas as pd
+import polars as pl
 import pytest
 import responses
 from pvlib.location import Location
@@ -80,14 +82,12 @@ class TestWeatherPlatform:
             raise ValueError(f"Fixture {request.param}_api_setup not found.")
 
     @pytest.fixture(params=[1, 2, 5, 10])
-    def max_forecast_day(self, request: pytest.FixtureRequest) -> pd.Timedelta:
-        return pd.Timedelta(days=request.param)
+    def max_forecast_day(self, request: pytest.FixtureRequest) -> dt.timedelta:
+        return dt.timedelta(days=request.param)
 
-    def convert_to_df(self, weather: dict[str, Any]) -> pd.DataFrame:
-        """Convert the weather data to a pd.DataFrame."""
-        weather_df: pd.DataFrame = pd.DataFrame.from_dict(weather["data"])
-        weather_df.set_index("datetime", inplace=True)
-        weather_df.index = pd.to_datetime(weather_df.index)
+    def convert_to_df(self, weather: dict[str, Any]) -> pl.DataFrame:
+        """Convert the weather data to a pl.DataFrame."""
+        weather_df: pl.DataFrame = pl.DataFrame(weather)
         return weather_df
 
     def test_weather_get_weather(self, weatherapi: WeatherAPI) -> None:
@@ -95,9 +95,10 @@ class TestWeatherPlatform:
         weather = weatherapi.get_weather()
         assert isinstance(weather, dict)
         weather = self.convert_to_df(weather)
-        assert isinstance(weather, pd.DataFrame)
-        # assert pd.infer_freq(weather.index) == "H"
-        assert not weather.isna().values.any()
+        assert isinstance(weather, pl.DataFrame)
+        # assert no value is NaN
+        sum_nan = weather.null_count().sum_horizontal()
+        assert sum_nan.item() == 0
         assert weather.shape[0] >= 24
 
     @pytest.mark.parametrize("freq", ["1H", "30Min", "15Min"])
@@ -108,15 +109,15 @@ class TestWeatherPlatform:
         weather = weatherapi.get_weather()
         assert isinstance(weather, dict)
         weather = self.convert_to_df(weather)
-        assert isinstance(weather, pd.DataFrame)
-        assert pd.infer_freq(weather.index) in time_aliases[freq]
+        assert isinstance(weather, pl.DataFrame)
+        assert pl.infer_freq(weather.index) in time_aliases[freq]
         assert not weather.isna().values.any()
         assert weather.shape[0] >= 24
 
     def test_weather_get_weather_max_days(
         self,
         weatherapi: WeatherAPI,
-        max_forecast_day: pd.Timedelta,
+        max_forecast_day: dt.timedelta,
         time_aliases: dict[str, list[str]],
     ) -> None:
         """Test the get_weather function with a number of higher data frequencies."""
@@ -125,11 +126,11 @@ class TestWeatherPlatform:
         weather = weatherapi.get_weather()
         assert isinstance(weather, dict)
         weather = self.convert_to_df(weather)
-        assert isinstance(weather, pd.DataFrame)
-        assert pd.infer_freq(weather.index) in time_aliases[freq]
+        assert isinstance(weather, pl.DataFrame)
+        assert pl.infer_freq(weather.index) in time_aliases[freq]
         assert not weather.isna().values.any()
         assert weather.shape[0] >= 24
-        assert weather.shape[0] <= max_forecast_day / pd.Timedelta(freq)
+        assert weather.shape[0] <= max_forecast_day / dt.timedelta(freq)
 
     def test_weather_data_cache(self, weatherapi: WeatherAPI) -> None:
         """Test the get_weather function."""
@@ -137,14 +138,14 @@ class TestWeatherPlatform:
         weather1 = weatherapi.get_weather()
         assert isinstance(weather1, dict)
         weather1 = self.convert_to_df(weather1)
-        assert isinstance(weather1, pd.DataFrame)
+        assert isinstance(weather1, pl.DataFrame)
         last_update1 = weatherapi._last_update
 
         # get second weather data object, should see that it is cached data
         weather2 = weatherapi.get_weather()
         assert isinstance(weather2, dict)
         weather2 = self.convert_to_df(weather2)
-        assert isinstance(weather2, pd.DataFrame)
+        assert isinstance(weather2, pl.DataFrame)
         last_update2 = weatherapi._last_update
         assert last_update1 == last_update2
 
@@ -155,13 +156,13 @@ class TestWeatherPlatform:
         weather1 = weatherapi.get_weather()
         assert isinstance(weather1, dict)
         weather1 = self.convert_to_df(weather1)
-        assert isinstance(weather1, pd.DataFrame)
+        assert isinstance(weather1, pl.DataFrame)
         last_update1 = weatherapi._last_update
 
         # get second weather data object, should see that it is live data
         weather2 = weatherapi.get_weather(live=True)
         assert isinstance(weather2, dict)
         weather2 = self.convert_to_df(weather2)
-        assert isinstance(weather2, pd.DataFrame)
+        assert isinstance(weather2, pl.DataFrame)
         last_update2 = weatherapi._last_update
         assert last_update1 != last_update2
